+++ src/lite/processor/parser.ts
@@ -4,18 +4,18 @@ function reducer(args: ClassValue[]) {
     return args.reduce<string[]>((acc, cur) => {
         if (!cur) return acc
         if (typeof cur === "object") {
-            Object.entries(cur).forEach(([key, values]) => {
+            for (const [key, values] of Object.entries(cur)) {
                 (Array.isArray(values) ? values.flat(Infinity) : [values]).forEach(
                     value => acc.push(
                         (parser[key] as (...args: ClassValue[]) => string)(value as ClassValue)
                     )
                 )
-            })
+            }
         } else {
             acc.push(...String(cur).split(" "))
         }
         return acc
-    }, []).flat()
+    }, []).flat().filter(Boolean)
 }
 
 export const parser = new Proxy((...args: ClassValue[]) => {
@@ -24,8 +24,8 @@ export const parser = new Proxy((...args: ClassValue[]) => {
     get: function (obj, key: string) {
         return key ? (
             ...args: ClassValue[]
-        ) => reducer(args).filter(values => values.trim() !== "").map((values) => (
-            values === "1" ? key : `${key}:${values}`.trim()
+        ) => reducer(args).map((values) => (
+            values === "NaN" ? key : `${key}:${values}`.trim()
         )) : obj
     }
 }) as Record<string, (...args: ClassValue[]) => string> & ((...args: ClassValue[]) => string)