+++ src/processor/parser.ts
@@ -10,18 +10,18 @@ function reducer(args: ClassValue[], options?: TWGOptions) {
     return args.reduce<string[]>((acc, cur) => {
         if (!cur) return acc
         if (typeof cur === "object") {
-            Object.entries(cur).forEach(([key, values]) => {
+            for (const [key, values] of Object.entries(cur)) {
                 (Array.isArray(values) ? values.flat(Infinity) : [values]).forEach(
                     value => acc.push(
                         (parser(options)[key] as (...args: ClassValue[]) => string)(value as ClassValue)
                     )
                 )
-            })
+            }
         } else {
             acc.push(...String(cur).split(" "))
         }
         return acc
-    }, []).flat()
+    }, []).flat().filter(Boolean)
 }
 
 /**
@@ -43,9 +43,9 @@ export function parser(options?: TWGOptions) {
         get: function (obj, key: string) {
             return key ? (
                 ...args: ClassValue[]
-            ) => reducer(args, options).filter(values => values.trim() !== "").map(
-                (values) => (values === "1" ? key : `${key}${divider}${values}`.trim())
-            ) : obj
+            ) => reducer(args, options).map((values) => (
+                values === "NaN" ? key : `${key}${divider}${values}`.trim()
+            )) : obj
         }
     }) as Record<string, (...args: ClassValue[]) => string> & ((...args: ClassValue[]) => string)
 }